Add poilcy families and build chai tests
