+++ include/RAJA/policy/PolicyBase.hpp
@@ -1,9 +1,13 @@
 #ifndef RAJA_POLICYBASE_HXX
 #define RAJA_POLICYBASE_HXX
 
+#include "RAJA/policy/PolicyFamily.hpp"
+
 namespace RAJA {
 
-struct PolicyBase { };
+struct PolicyBase { 
+  const PolicyFamily family = PolicyFamily::undefined;
+};
 
 }
