+++ include/RAJA/policy/openmp/policy_openmp.hpp
@@ -1,6 +1,8 @@
 #ifndef policy_openmp_HXX_
 #define policy_openmp_HXX_
 
+#include "RAJA/policy/PolicyFamily.hpp"
+
 namespace RAJA
 {
 
@@ -12,24 +14,30 @@ namespace RAJA
 //////////////////////////////////////////////////////////////////////
 //
 
+struct omp_exec_base {};
+
 ///
 /// Segment execution policies
 ///
 template <typename InnerPolicy>
-struct omp_parallel_exec {
+struct omp_parallel_exec  : public omp_exec_base {
+  const PolicyFamily family = PolicyFamily::openmp;
 };
-struct omp_for_exec {
+struct omp_for_exec : public omp_exec_base { 
+  const PolicyFamily family = PolicyFamily::openmp;
 };
 struct omp_parallel_for_exec : public omp_parallel_exec<omp_for_exec> {
 };
 template <size_t ChunkSize>
-struct omp_for_static {
+struct omp_for_static : public omp_exec_base {
+  const PolicyFamily family = PolicyFamily::openmp;
 };
 template <size_t ChunkSize>
 struct omp_parallel_for_static
     : public omp_parallel_exec<omp_for_static<ChunkSize>> {
 };
-struct omp_for_nowait_exec {
+struct omp_for_nowait_exec : public omp_exec_base {
+  const PolicyFamily family = PolicyFamily::openmp;
 };
 
 ///
@@ -39,9 +47,11 @@ struct omp_parallel_for_segit : public omp_parallel_for_exec {
 };
 struct omp_parallel_segit : public omp_parallel_for_segit {
 };
-struct omp_taskgraph_segit {
+struct omp_taskgraph_segit : public omp_exec_base {
+  const PolicyFamily family = PolicyFamily::openmp;
 };
-struct omp_taskgraph_interval_segit {
+struct omp_taskgraph_interval_segit : public omp_exec_base {
+  const PolicyFamily family = PolicyFamily::openmp;
 };
 
 ///
