+++ include/RAJA/policy/simd/policy_simd.hpp
@@ -1,6 +1,9 @@
 #ifndef policy_simd_HXX_
 #define policy_simd_HXX_
 
+#include "RAJA/policy/PolicyBase.hpp"
+#include "RAJA/policy/PolicyFamily.hpp"
+
 //
 //////////////////////////////////////////////////////////////////////
 //
@@ -15,7 +18,8 @@
 namespace RAJA
 {
 
-struct simd_exec {
+struct simd_exec : public PolicyBase {
+  const PolicyFamily family = PolicyFamily::simd;
 };
 
 } // end of namespace RAJA
