+++ include/RAJA/util/chai_support.hpp
@@ -2,19 +2,15 @@
 #define RAJA_DETAIL_RAJA_CHAI_HPP
 
 #include "chai/ExecutionSpaces.hpp"
+#include "RAJA/policy/PolicyBase.hpp"
+#include "RAJA/index/IndexSet.hpp"
+#include "RAJA/internal/LegacyCompatibility.hpp"
+#include "RAJA/internal/ForallNPolicy.hpp"
 
 #include "RAJA/policy/sequential/policy_sequential.hpp"
 #include "RAJA/policy/simd/policy_simd.hpp"
 #include "RAJA/policy/cuda/policy_cuda.hpp"
 #include "RAJA/policy/openmp/policy_openmp.hpp"
-
-#include "RAJA/policy/PolicyBase.hpp"
-
-#include "RAJA/index/IndexSet.hpp"
-
-#include "RAJA/internal/LegacyCompatibility.hpp"
-
-#include "RAJA/internal/ForallNPolicy.hpp"
 
 #include <tuple>
 #include <type_traits>
@@ -23,39 +19,24 @@
 namespace RAJA {
 namespace detail {
 
-constexpr chai::ExecutionSpace getSpace(const ::RAJA::PolicyBase) {
-  return chai::CPU;
+const chai::ExecutionSpace getSpace(const ::RAJA::PolicyBase) {
+  return ::chai::CPU;
 }
 
 constexpr chai::ExecutionSpace getSpace(const ::RAJA::omp_exec_base) {
-  return chai::CPU;
+  return ::chai::CPU;
 }
 
 constexpr chai::ExecutionSpace getSpace(const ::RAJA::cuda_exec_base) {
-  return chai::GPU;
+  return ::chai::GPU;
 }
 
 template <typename A, typename B>
 constexpr chai::ExecutionSpace getSpace(const ::RAJA::IndexSet::ExecPolicy<A,B>) {
-  return chai::NONE;
+  return ::chai::NONE;
 }
 
 #if defined(RAJA_ENABLE_NESTED)
-template <typename T, typename Tuple>
-struct has_type;
-
-template <typename T>
-struct has_type<T, std::tuple<>> : std::false_type {};
-
-template <typename T, typename U, typename... Ts>
-struct has_type<T, std::tuple<U, Ts...>> : has_type<T, std::tuple<Ts...>> {};
-
-template <typename T, typename... Ts>
-struct has_type<T, std::tuple<T, Ts...>> : std::true_type {};
-
-template <typename T, typename Tuple>
-using tuple_contains_type = typename has_type<T, Tuple>::type;
-
 template <typename policy>
 int is_cuda (policy p) {
   if (p.family == cuda)
